// src/app/models/book.model.ts
export class Book{
  isbn: string;
  subtitle: string;
  published: string;
  publisher: string;
  description: string;
  category: string;
  availability: string;
  author: string;
  country: string;
  imageLink: string;
  language: string;
  website: string;
  pages: number;
  title: string;
  year: number
}